% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{pyany_auth}
\alias{pyany_auth}
\title{Set credentials to pythonanywhere as environment variables}
\usage{
pyany_auth(username, token)
}
\arguments{
\item{username}{Your pythonanywhere username}

\item{token}{Your pythonanywhere personal token}
}
\value{
sets envs
}
\description{
Initialize env variables. Note that they disappear after session is closed
}
\examples{
pyany_auth(username, token)

}
