% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{pyany_download_file}
\alias{pyany_download_file}
\title{Download a file}
\usage{
pyany_download_file(file_path, file_extension = NULL)
}
\arguments{
\item{file_path}{Path to file, for example folder_name/file_name.extension}

\item{file_extension}{File extension. For now only .rds and .csv are working}
}
\value{
content of csv (tibble) or rds (tibble or other objects)
}
\description{
Downloads a file
}
\examples{
an_object <- pyany_download_file(file_path, file_extension = "csv")

}
