% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{pyany_list_files}
\alias{pyany_list_files}
\title{List content in given folder}
\usage{
pyany_list_files(folder_name)
}
\arguments{
\item{folder_name}{Directory name as displayed in files cart}
}
\value{
object of class pythonanywhere_api
}
\description{
List all files in directory
}
\examples{
pyany_list_files(folder_name)

}
